move build-code/code -> build-code/sync
